consolidate aggregation, multitenancy and multientities
